Movidas las actualizaciones a app_2
