+++ GoBurguer/settings.py
@@ -37,7 +37,8 @@ INSTALLED_APPS = [
     'django.contrib.sessions',
     'django.contrib.messages',
     'django.contrib.staticfiles',
-    'app'
+    'app',
+    'app_2'
 ]
 
 MIDDLEWARE = [
