+++ app_2/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class App2Config(AppConfig):
+    default_auto_field = 'django.db.models.BigAutoField'
+    name = 'app_2'
