+++ app_2/static/carrito.css
@@ -1,0 +1,23 @@
+button {
+    background-color: #ffc107;
+    color: #000;
+    padding: 12px 25px;
+    font-size: 16px;
+    border: none;
+    border-radius: 8px;
+    cursor: pointer;
+    transition: background-color 0.3s ease;
+    margin-top: 20px;
+}
+
+.button-group {
+    display: flex;
+    justify-content: space-between;
+    gap: 20px; /* Añade espacio entre los botones */
+    width: 100%;
+}
+
+button:hover {
+    background-color: #ffca2c;
+}
+
