+++ app_2/static/ingredientes.css
@@ -1,0 +1,99 @@
+body {
+    font-family: 'Poppins', sans-serif;
+    background: #ffffff; /* Fondo blanco */
+    margin: 0;
+    padding: 0;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    min-height: 100vh;
+}
+
+h1 {
+    text-align: center;
+    color: #000000; /* Texto en negro */
+    margin-bottom: 20px;
+    font-size: 32px;
+    font-weight: bold;
+}
+
+.ingredientes-section {
+    background-color: #f8f8f8; /* Fondo gris claro para la sección de ingredientes */
+    padding: 30px;
+    border-radius: 15px;
+    box-shadow: 0px 8px 16px rgba(0, 0, 0, 0.1);
+    max-width: 900px;
+    width: 100%;
+}
+
+.ingredientes-list {
+    list-style-type: none;
+    padding: 0;
+    margin: 0;
+}
+
+.ingrediente-item {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    background-color: #fff; /* Fondo blanco para cada ingrediente */
+    padding: 15px;
+    border-radius: 12px;
+    margin-bottom: 10px;
+    transition: transform 0.2s ease-in-out;
+}
+
+.ingrediente-item:hover {
+    transform: scale(1.03);
+    background-color: #ffcc00; /* Efecto hover gris claro */
+}
+
+.ingrediente-img {
+    width: 60px;
+    height: 60px;
+    object-fit: cover;
+    border-radius: 50%;
+    margin-right: 20px;
+}
+
+.ingredient-info {
+    flex-grow: 1;
+    display: flex;
+    flex-direction: column;
+}
+
+.ingredient-info span {
+    font-size: 18px;
+    font-weight: 600;
+    color: #000000; /* Texto de ingredientes en negro */
+}
+
+.ingredient-checkbox {
+    margin-left: 20px;
+    transform: scale(1.2);
+}
+
+p {
+    text-align: right;
+    font-size: 18px;
+    font-weight: bold;
+    margin-top: 20px;
+    color: #000000; /* Texto del total en amarillo */
+}
+
+button {
+    background-color: #ffcc00; /* Botón en amarillo */
+    color: #000000; /* Texto del botón en negro */
+    padding: 12px 25px;
+    font-size: 16px;
+    border: none;
+    border-radius: 8px;
+    cursor: pointer;
+    width: 100%;
+    transition: background-color 0.3s ease;
+    margin-top: 20px;
+}
+
+button:hover {
+    background-color: #e6b800; /* Efecto hover en amarillo más oscuro */
+}
