+++ app_2/templates/verificar-reset-code.html
@@ -1,0 +1,34 @@
+<!DOCTYPE html>
+<html lang="es">
+<head>
+    {% load static %}
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Verificar codigo</title>
+    <link rel="stylesheet" href="{% static 'register.css' %}"> <!-- Reutilizamos el CSS de registro -->
+</head>
+<body>
+    <div class="container">
+        <!-- Sección izquierda con una imagen -->
+        <div class="left-section">
+            <img src="{% static 'img/9343597.jpg' %}" alt="Ilustración" class="illustration">
+        </div>
+
+        <!-- Sección derecha con el formulario de inicio de sesión -->
+        <div class="right-section">
+            <div class="form-container">
+                <h2>Ingresar codigo de verificación</h2>
+                    <form method="POST">
+                        {% csrf_token %}
+                        {{ form.as_p }}
+                        <button type="submit" class="btn">Verificar</button>
+                    </form>
+            </div>
+        </div>
+    </div>
+
+    <footer>
+        <p>© 2024 GoyoBurger - Todos los derechos reservados</p>
+    </footer>
+</body>
+</html>
